Move the OS-identification code into NativeLibraryLoader
This makes it easier to replace in scenarios where the user wants
to load libraries differently, since it doesn't require any changes
to Platform, it just requires replacing NativeCryptoJni and
NativeLibraryLoader wholesale.
